hide password/configure user rights/install cors

+++ controllers/clothingItems.js
@@ -33,22 +33,26 @@ const createClothingItem = (req, res) => {
 
 const deleteClothingItem = (req, res) => {
   const { itemId } = req.params;
-  clothingItem
-    .findByIdAndDelete(itemId)
-    .orFail()
-    .then((item) => res.send(item))
-    .catch((err) => {
-      console.error(err);
-      if (err.name === "DocumentNotFoundError") {
-        return res.status(NOT_FOUND).send({ message: err.message });
-      }
-      if (err.name === "CastError") {
-        return res.status(BAD_REQUEST).send({ message: "Invalid data" });
-      }
-      return res
-        .status(DEFAULT)
-        .send({ message: "An error has occured on the server." });
-    });
+  console.log(req.user);
+  if (req.user._id === itemId) {
+    clothingItem
+      .findByIdAndDelete(itemId)
+      .orFail()
+      .then((item) => res.send(item))
+      .catch((err) => {
+        console.error(err);
+        if (err.name === "DocumentNotFoundError") {
+          return res.status(NOT_FOUND).send({ message: err.message });
+        }
+        if (err.name === "CastError") {
+          return res.status(BAD_REQUEST).send({ message: "Invalid data" });
+        }
+        return res
+          .status(DEFAULT)
+          .send({ message: "An error has occured on the server." });
+      });
+  }
+  return res.status(403).send({ message: "Cannot delete item" });
 };
 
 const addLike = (req, res) => {
